+++ lib/occams-record/eager_loaders/context.rb
@@ -77,10 +77,11 @@ module OccamsRecord
       # @param as [Symbol] Load the association usign a different attribute name
       # @param from [Symbol] Opposite of `as`. `assoc` is the custom name and `from` is the name of association on the ActiveRecord model.
       # @param optimizer [Symbol] Only used for `through` associations. Options are :none (load all intermediate records) | :select (load all intermediate records but only SELECT the necessary columns)
+      # @param active_record_fallback [Symbol] If passed, missing methods will be forwarded to an ActiveRecord instance. Options are :lazy (allow lazy loading in the AR record) or :strict (require strict loading)
       # @yield a block where you may perform eager loading on *this* association (optional)
       # @return [OccamsRecord::EagerLoaders::Base] the new loader. if @model is nil, nil will be returned.
       #
-      def add(assoc, scope = nil, select: nil, use: nil, as: nil, from: nil, optimizer: :select, &builder)
+      def add(assoc, scope = nil, select: nil, use: nil, as: nil, from: nil, optimizer: :select, active_record_fallback: nil, &builder)
         if from
           real_assoc = from
           custom_name = assoc
@@ -93,11 +94,11 @@ module OccamsRecord
         end
 
         if @model
-          loader = build_loader!(real_assoc, custom_name, scope, select, use, optimizer, builder)
+          loader = build_loader!(real_assoc, custom_name, scope, select, use, optimizer, builder, active_record_fallback)
           @loaders << loader
           loader
         else
-          @dynamic_loaders << [real_assoc, custom_name, scope, select, use, optimizer, builder]
+          @dynamic_loaders << [real_assoc, custom_name, scope, select, use, optimizer, builder, active_record_fallback]
           nil
         end
       end
@@ -116,6 +117,7 @@ module OccamsRecord
         nil
       end
 
+      # Iterate over each loader
       def each
         return @loaders.each unless block_given?
         @loaders.each { |l| yield l }
@@ -123,19 +125,19 @@ module OccamsRecord
 
       private
 
-      def build_loader!(assoc, custom_name, scope, select, use, optimizer, builder)
-        build_loader(assoc, custom_name, scope, select, use, optimizer, builder) ||
+      def build_loader!(assoc, custom_name, scope, select, use, optimizer, builder, active_record_fallback)
+        build_loader(assoc, custom_name, scope, select, use, optimizer, builder, active_record_fallback) ||
           raise("OccamsRecord: No association `:#{assoc}` on `#{@model.name}` or subclasses")
       end
 
-      def build_loader(assoc, custom_name, scope, select, use, optimizer, builder)
+      def build_loader(assoc, custom_name, scope, select, use, optimizer, builder, active_record_fallback)
         ref = @model.reflections[assoc.to_s] ||
           @model.subclasses.map(&:reflections).detect { |x| x.has_key? assoc.to_s }&.[](assoc.to_s)
         return nil if ref.nil?
 
         scope ||= ->(q) { q.select select } if select
         loader_class = !!ref.through_reflection ? EagerLoaders::Through : EagerLoaders.fetch!(ref)
-        loader_class.new(ref, scope, use: use, as: custom_name, optimizer: optimizer, parent: @tracer, &builder)
+        loader_class.new(ref, scope, use: use, as: custom_name, optimizer: optimizer, parent: @tracer, active_record_fallback: active_record_fallback, &builder)
       end
     end
   end