+++ lib/occams-record/eager_loaders/through.rb
@@ -9,7 +9,7 @@ module OccamsRecord
       #
       # See documentation for OccamsRecord::EagerLoaders::Base.
       #
-      def initialize(ref, scope = nil, use: nil, as: nil, optimizer: :select, parent: nil, &builder)
+      def initialize(ref, scope = nil, use: nil, as: nil, optimizer: :select, parent: nil, active_record_fallback: nil, &builder)
         super
 
         unless @ref.macro == :has_one or @ref.macro == :has_many
@@ -83,7 +83,7 @@ module OccamsRecord
             nested_loader.tracer.through = true
             nested_loader
           }.
-          nest(tail.ref.source_reflection.name, @scope, use: @use, as: @as)
+          nest(tail.ref.source_reflection.name, @scope, use: @use, as: @as, active_record_fallback: @active_record_fallback)
 
         @eager_loaders.each { |loader| inner_loader.eager_loaders << loader }
         inner_loader.tracer.name = tracer.name