+++ lib/occams-record/results/results.rb
@@ -17,10 +17,14 @@ module OccamsRecord
     # @param association_names [Array<String>] names of associations that will be eager loaded into the results.
     # @param model [ActiveRecord::Base] the AR model representing the table (it holds column & type info).
     # @param modules [Array<Module>] (optional)
-    # @param [OccamsRecord::EagerLoaders::Base] the eager loaded that loaded this class of records
+    # @param tracer [OccamsRecord::EagerLoaders::Tracer] the eager loaded that loaded this class of records
+    # @param active_record_fallback [Symbol] If passed, missing methods will be forwarded to an ActiveRecord instance. Options are :lazy (allow lazy loading in the AR record) or :strict (require strict loading)
     # @return [OccamsRecord::Results::Row] a class customized for this result set
     #
-    def self.klass(column_names, column_types, association_names = [], model: nil, modules: nil, tracer: nil)
+    def self.klass(column_names, column_types, association_names = [], model: nil, modules: nil, tracer: nil, active_record_fallback: nil)
+      raise ArgumentError, "Invalid active_record_fallback option :#{active_record_fallback}. Valid options are :lazy, :strict" if active_record_fallback and !%i(lazy strict).include?(active_record_fallback)
+      raise ArgumentError, "Option active_record_fallback is not allowed when no model is present" if active_record_fallback and model.nil?
+
       Class.new(Results::Row) do
         Array(modules).each { |mod| prepend mod } if modules
 
@@ -30,6 +34,7 @@ module OccamsRecord
         self.model_name = model ? model.name : nil
         self.table_name = model ? model.table_name : nil
         self.eager_loader_trace = tracer
+        self.active_record_fallback = active_record_fallback
         self.primary_key = if model&.primary_key and (pkey = model.primary_key.to_s) and columns.include?(pkey)
                              pkey
                            end