+++ HISTORY.md
@@ -1,6 +1,7 @@
 ### 1.6.2 (2023-02-17)
 * Bugfix for eager loading inside a 'through' association
 * Improve traces for 'through' associations
+* Add the `active_record_fallback` option to `eager_load` to allow missing methods to automatically be forwarded to an ActiveRecord instance.
 
 ### 1.6.1 (2023-02-17)
 * Add an 'eager load trace' to missing association/column errors. Otherwise, in a very large set of eager loads, it can be virtually impossible to tell where the missing one is.